+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 English | [简体中文](./CHANGELOG_CN.md)
 
+## 3.12.0 (2022-02-17)
+
+- `Feat(Core)` Add new static property `VConsole.instance` to get the singleton instance.
+- `Feat(Core)` Add new options `storage.defaultStorages`, see [Public Properties & Methods](./doc/public_properties_methods.md).
+- `Feat(Core)` New way of using `vConsole.setOption()`: `setOption('log.maxLogNumber', 20)` to set `maxLogNumber` field only, and `setOption({ log: { maxLogNumber: 20 }})` to overwrite `log` object.
+- `Feat(Core)` Deprecated some options, see below.
+- `Fix(Plugin)` Fix the bug that event `renderTab` doesn't render plugin view.
+- `Fix(Storage)` Fix cookie parse error in some bad cases. (issue #508, #509)
+
+**Deprecated Options:**
+
+- `maxLogNumber`: Use `option.log.maxLogNumber` instead.
+- `maxNetworkNumber`: Use `option.network.maxNetworkNumber` instead.
+- `onClearLog`: Removed.
+
+
 ## 3.11.2 (2022-01-20)
 
 - `Feat(Storage)` Added "Clear" button to batch delete all storage items. (issue #499)